+++ dyn-cache/CHANGELOG.md
@@ -12,6 +12,7 @@ invocations.
 - Crate extracted from `topo::cache` module.
 - `{LocalCache,SendCache}::cache` wraps `cache_with` for types that impl `Clone`.
 - `Gc` trait is public, inner cache types implement it.
+- `Hashed` has an additional type parameter for the type of hasher used to create it.
 
 ### Changed